+++ changelog.md
@@ -25,6 +25,9 @@
   * Added fish swim speed config option (defaults to 2.5x vanilla speed)
 * Snow Golems Bonus Stats has been moved to MPR Data Pack
   * The feature also made snowballs not trigger invulnerability frames, so it's now a "new" feature
+* Beta Strafe
+  * Enhanced calculations of speed and angle
+  * Now makes mobs jump when colliding into something instead of checking blocks in the direction of the strafing
 * Only 50% of mobs will now get the xray range override
 * Fisher Mobs
   * Fishing range is now configurable